Updated to the 2021 library

+++ src/main/java/frc/robot/subsystems/Chassis.java
@@ -257,7 +257,7 @@ public void stopDriveMotors() {
     return Math.abs(leftLead.getClosedLoopError()) < 1000;
   }
 
-  public int getError() {
+  public double getError() {
     return leftLead.getClosedLoopError();
   }
